Merged PR 2055: Move to Spacy v2.0 and fix multi-word place names
Spacy v2.0 is faster and more flexible.
The API is a bit changed, so the fixes to bug 2933 are on latest Spacy code.
Download of models is now done by spying or injecting into PIP with the models we want. They are a LOT smaller than before.
We download multiple models, but the server is only loading the existing EN models as before for now.

The issue with 2933 is that "New York" and "York" clash and we have to merge the best entity candidate. In this case we merge in the longest entity by token (word). We assume that for a similar length token our custom match will win. The results of clashes are logged.

Rather than adding a sentence case and lower case form of the entity, we match to LOWER. Also the cities list is processed to remove duplicate entities that would cause excess noise in the entity merge algorithm.

Add an exception wrapper for the aiohttp, so we can't fail to log exceptions during the processing of a web call.

Related work items: #2933